+++ src/lib/components/document/Document.tsx
@@ -468,6 +468,7 @@ const Document = ({ data }: IDocumentProps) => {
     try {
       setEditStage(EDIT_STAGES.WORKING)
       const svg = document.getElementById('evilFlowersPaintSVG')!
+      svg.classList.remove('efw-border', 'efw-border-red') // remove border before save/update
       if (layer)
         await editPackage!.updateLayerFunc(layer.id, svg, groupId, activePage)
       else {